QuickFix
Made it so that the enemy will look at the player, so when shooting it is always shooting in the right direction. Also, changed the shooting enemy so that it does not go up to the player, it leaves a little gap to continue shooting if the player is standing still.